requirements.txt
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
Deprecated==1.2.5
idna==2.8
PyGithub==1.43.7
PyJWT==1.7.1
requests==2.22.0
soupsieve==1.9.1
urllib3==1.25.3
wrapt==1.11.2
